Reusable components moved to common folder
