+++ js/components/developer/common/searchListStyles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+const searchListStyles = StyleSheet.create({
+  searchListView: {
+    flex: 1,
+  },
+  searchListCardItem: {
+    alignItems: 'flex-start',
+    flex: 1,
+  },
+});
+
+module.exports = searchListStyles;
