+++ js/components/developer/language-picker-screen/languageDisplayStyles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { TITLE_COLOR, MAIN_CERISE, SECONDARY_GRAY } from './commonColors';
+import { TITLE_COLOR, MAIN_CERISE, SECONDARY_GRAY } from '../common/colors';
 
 const languageDisplayStyles = StyleSheet.create({
   rowList: {
