--- js/components/developer/language-picker-screen/languageSelectionStyles.js
@@ -1,14 +1,0 @@
-import { StyleSheet } from 'react-native';
-// import { TITLE_COLOR, MAIN_CERISE, SECONDARY_GRAY } from './commonColors';
-
-const languageSelectionStyles = StyleSheet.create({
-  languageSelectionView: {
-    flex: 1,
-  },
-  languageSelectionList: {
-    alignItems: 'flex-start',
-    flex: 1,
-  },
-});
-
-module.exports = languageSelectionStyles;
